works :)

+++ engine/src/simulation/world/detail/rule/action/calculation/Calculation.java
@@ -7,10 +7,10 @@ import simulation.world.detail.rule.action.Action;
 
 public abstract class Calculation extends Action {
     EntityProperty resultProp = null;
-    double arg1 = 0;
-    double arg2 = 0;
+    float arg1 = 0;
+    float arg2 = 0;
 
-    public Calculation(Entity entity, String propertyName, double arg1, double arg2) {
+    public Calculation(Entity entity, String propertyName, float arg1, float arg2) {
         super(entity, propertyName);
         this.arg1 = arg1;
         this.arg2 = arg2;
