+++ engine/src/simulation/world/detail/rule/action/calculation/Divide.java
@@ -1,14 +1,10 @@
 package simulation.world.detail.rule.action.calculation;
 
-import simulation.utils.Value;
 import simulation.world.detail.entity.Entity;
 import simulation.world.detail.entity.EntityInstance;
-import simulation.world.detail.rule.action.Action;
-
-import static java.lang.Double.parseDouble;
 
 public class Divide extends Calculation {
-    public Divide(Entity entity, String propertyName, double arg1, double arg2) {
+    public Divide(Entity entity, String propertyName, float arg1, float arg2) {
         super(entity, propertyName, arg1, arg2);
     }
